refactor: 메소드 위치 이동
